requirements.txt
flask
rfc3339
requests
flask_sqlalchemy
flask_migrate
flask_shell_ipython
flask_marshmallow
flask_cors
gitpython
marshmallow_sqlalchemy
boto3
balena-sdk
simplejson
pytest
pytest-pudb
pytest-cov
celery
pytest-celery
supervisor
python-dotenv
mysqlclient
